finish

+++ app/Http/Controllers/PostsController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
 use App\Post;
+use Illuminate\Support\Facades\Storage;
 
 class PostsController extends Controller
 {
@@ -67,8 +68,28 @@ class PostsController extends Controller
     {
           $this->validate($request,[
              'title' => 'required',
-             'body'=> 'required'
+             'body'=> 'required',
+              'cover_image' => 'image|nullable|max:1999'
           ]);
+
+       // Hadle file upload
+        if($request->hasFile('cover_image')){
+
+            // get filename with extension
+
+             $fileNameWithExt = $request->file('cover_image')->getClientOriginalName();
+            // Get just filename
+             $filename = pathinfo($fileNameWithExt,PATHINFO_FILENAME);
+            //get just ext
+              $extension = $request->file('cover_image')->getClientOriginalExtension();
+
+             $fileNameToStore = $filename.'_'.time().'.'.$extension;
+
+            $path = $request->file('cover_image')->storeAs('public/cover_images',$fileNameToStore);
+        }
+        else{
+            $fileNameToStore = 'noimage.jpg';
+        }
 
        // Create post
 
@@ -80,6 +101,7 @@ class PostsController extends Controller
        $post->title = $request->input('title');
        $post->body = $request->input('body');
        $post->user_id =auth()->user()->id;
+       $post->cover_image = $fileNameToStore;
        $post->save();
 
        return redirect('/posts')->with('success','Post Created');
@@ -128,6 +150,22 @@ class PostsController extends Controller
             'body'=> 'required'
          ]);
 
+        // Hadle file upload
+        if($request->hasFile('cover_image')){
+
+            // get filename with extension
+
+            $fileNameWithExt = $request->file('cover_image')->getClientOriginalName();
+            // Get just filename
+            $filename = pathinfo($fileNameWithExt,PATHINFO_FILENAME);
+            //get just ext
+            $extension = $request->file('cover_image')->getClientOriginalExtension();
+
+            $fileNameToStore = $filename.'_'.time().'.'.$extension;
+
+            $path = $request->file('cover_image')->storeAs('public/cover_images',$fileNameToStore);
+        }
+
       // Create post
 
        /*
@@ -137,6 +175,9 @@ class PostsController extends Controller
       $post = Post::find($id);
       $post->title = $request->input('title');
       $post->body = $request->input('body');
+      if($request->hasFile('cover_image')){
+          $post->cover_image = $fileNameToStore;
+      }
       $post->save();
 
       return redirect('/posts')->with('success','Post Updated');    }
@@ -150,6 +191,15 @@ class PostsController extends Controller
     public function destroy($id)
     {
         $post = Post::find($id);
+        if(auth()->user()->id !== $post->user_id){
+            return redirect('/posts')->with('error','Unathorized Page');
+        }
+
+        if($post->cover_image != 'noimage.jpg'){
+            // DELETE IMAGE
+            Storage::delete('public/cover_images/'.$post->cover_image);
+        }
+
         $post->delete();
         return redirect('/posts')->with('success','Post Removed');
     }
